[MOVE] ia part in a file
